% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVM.test.R
\name{CVM.test}
\alias{CVM.test}
\title{Two-sample Cramer-von Mises test}
\usage{
CVM.test(x, y, R)
}
\arguments{
\item{x}{independent random samples from distribution F}

\item{y}{independent random samples from distribution G}

\item{R}{number of repeated sampling}
}
\value{
the p-value of hypothesis test
}
\description{
Choose Cramer-von Mises statistic to measure the difference between two distributions and test for equal distributions in the univariate case
}
\examples{
\dontrun{
attach(chickwts)
x <- sort(as.vector(weight[feed == "soybean"]))
y <- sort(as.vector(weight[feed == "linseed"]))
detach(chickwts)
CVM.test(x,y,R=999)
}
}
