% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my.chisq.test.R
\name{my.chisq.test}
\alias{my.chisq.test}
\title{A faster version of chisq.test()}
\usage{
my.chisq.test(x, y)
}
\arguments{
\item{x}{a numeric vector with no missing values}

\item{y}{a numeric vector with no missing values}
}
\value{
the chi-square test statistic
}
\description{
A faster version of chisq.test() that only computes the chi-square test statistic when the input is two numeric vectors with no missing values.
}
\examples{
\dontrun{
my.chisq.test(c(367,342,266,329),c(56,40,20,16))
}
}
