% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my.pcauchy.R
\name{my.pcauchy}
\alias{my.pcauchy}
\title{Compute the cdf of the Cauchy distribution}
\usage{
my.pcauchy(x, eta, theta)
}
\arguments{
\item{x}{the upper bound}

\item{eta}{the location parameter of the Cauchy distribution}

\item{theta}{the scale parameter of the Cauchy distribution}
}
\value{
the value of the cdf of the Cauchy distribution at x
}
\description{
Compute the cdf of the Cauchy distribution
}
\examples{
\dontrun{
my.pcauchy(x=1, eta=0, theta=1)
}
}
