% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my.table.R
\name{my.table}
\alias{my.table}
\title{A faster version of table()}
\usage{
my.table(x, y)
}
\arguments{
\item{x}{an integer vector with no missing values}

\item{y}{an integer vector with no missing values}
}
\value{
a contingency table
}
\description{
A faster version of table() for the case of an input of two integer vectors with no missing values
}
\examples{
\dontrun{
x <- c(1, 2, 2, 3, 1, 3, 3, 2)
y <- c(1, 1, 2, 1, 1, 1, 1, 2)
my.table(x, y)
}
}
